Storm grenade is functional. Dodge cooldown display is functional.
Dodge cooldown changes to black when out of light (cannot recharge) and green when in light (can recharge). Fades in when not at max charge and fades out at max charge. The cooldown does not deplete during dodges or telepoting.
Storm grenade blocks light such that the player is considered out of light and the player can teleport into it.